Test case for handling single bracket done

+++ src/kajaljad/expressionevaluator/lib/Operations.java
@@ -4,7 +4,7 @@ import java.util.HashMap;
 import java.util.Map;
 
 interface Evaluate {
-    int evaluate(int number1, int number2);
+    double evaluate(double number1, double number2);
 }
 public class Operations {
     Map<String,Evaluate> operationMap;
@@ -14,37 +14,37 @@ public class Operations {
 
         map.put("+", new Evaluate() {
             @Override
-            public int evaluate(int number1, int number2) {
+            public double evaluate(double number1, double number2) {
                 return number1 + number2;
             }
         });
         map.put("-", new Evaluate() {
             @Override
-            public int evaluate(int number1, int number2) {
+            public double evaluate(double number1, double number2) {
                 return number1 - number2;
             }
         });
         map.put("*", new Evaluate() {
             @Override
-            public int evaluate(int number1, int number2) {
+            public double evaluate(double number1, double number2) {
                 return number1 * number2;
             }
         });
         map.put("/", new Evaluate() {
             @Override
-            public int evaluate(int number1, int number2) {
+            public double evaluate(double number1, double number2) {
                 return number1 / number2;
             }
         });
         map.put("^", new Evaluate() {
             @Override
-            public int evaluate(int number1, int number2) {
-                return (int)Math.pow(number1, number2);
+            public double evaluate(double number1, double number2) {
+                return Math.pow(number1, number2);
             }
         });
         this.operationMap = map;
     }
-    public int performOperations(String operator, int number1, int number2) {
+    public double performOperations(String operator, double number1, Double number2) {
         return operationMap.get(operator).evaluate(number1,number2);
     }
